requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0

# Scheduling
schedule>=1.2.0

# AI Processing
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Web Framework
flask>=3.0.0

# Database
sqlite3  # Built-in with Python

# Utilities
python-dotenv>=1.0.0
hashlib  # Built-in with Python
json  # Built-in with Python
datetime  # Built-in with Python
logging  # Built-in with Python
threading  # Built-in with Python
time  # Built-in with Python
os  # Built-in with Python
re  # Built-in with Python
urllib  # Built-in with Python

# Optional: Advanced features
scikit-learn>=1.3.0  # For advanced duplicate detection
psutil>=5.9.0  # For system monitoring
prometheus-client>=0.19.0  # For metrics collection
grafana-api>=1.0.3  # For Grafana integration

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
